+++ source/drivers/udma/v0/hw_include/csl_lcdma_ringacc.c
@@ -181,7 +181,7 @@ static int32_t CSL_lcdma_ringaccPush64MultiAccess( CSL_LcdmaRingaccCfg *pCfg, CS
              *---------------------------------------------------------------*/
             if( (pfMemOps != NULL) && (pRing->asel == (uint32_t)0U) )
             {
-                (*pfMemOps)((void *)pRingEntry, numValuesWritten * sizeof(uint64_t), CSL_LCDMA_RINGACC_MEM_OPS_TYPE_WR);
+                (*pfMemOps)((void *)pRingEntry, (uint32_t)(numValuesWritten * sizeof(uint64_t)), CSL_LCDMA_RINGACC_MEM_OPS_TYPE_WR);
             }
             CSL_lcdma_ringaccCommitToForwardRing( pCfg, pRing, (int32_t)numValuesWritten );    /* This call will update wrOcc and wrIdx elements in pRing */
             if( numValues != numValuesWritten )
@@ -228,7 +228,7 @@ static int32_t CSL_lcdma_ringaccPop64MultiAccess( CSL_LcdmaRingaccCfg *pCfg, CSL
             if( (pfMemOps != NULL) && (pRing->asel == (uint32_t)0U) )
             {
                 pRingEntry = (void *)(((uintptr_t)localRdIdx * pRing->elSz) + (uintptr_t)pRing->virtBase);
-                (*pfMemOps)((void *)pRingEntry, numValuesRead * sizeof(uint64_t), CSL_LCDMA_RINGACC_MEM_OPS_TYPE_RD);
+                (*pfMemOps)((void *)pRingEntry, (uint32_t)(numValuesRead * sizeof(uint64_t)), CSL_LCDMA_RINGACC_MEM_OPS_TYPE_RD);
             }
             for( i=0U; i<numValuesRead; i++ )
             {
