+++ source/drivers/udma/v0/soc/am62ax/udma_soc.c
@@ -286,10 +286,10 @@ uint32_t Udma_isCacheCoherent(void)
 
 uint8_t Udma_isValidInstance(uint32_t instId)
 {
-    uint8_t result = FALSE;
+    uint8_t result = (uint8_t)FALSE;
     if((UDMA_INST_ID_BCDMA_0 == instId) || (UDMA_INST_ID_PKTDMA_0 == instId))
     {
-        result = TRUE;
+        result = (uint8_t)TRUE;
     }
 
     return result;
